Split the functions into multiple files
