+++ deer/__init__.py
@@ -1,1 +1,3 @@
-from deer.seq1d import *
+from deer.fseq1d import *
+from deer.fsolve_idae import *
+from deer.fsolve_ivp import *
