+++ deer/maths.py
@@ -1,0 +1,46 @@
+from typing import Tuple
+import jax
+import jax.numpy as jnp
+
+
+def scan_binop(element_i: Tuple[jnp.ndarray, jnp.ndarray],
+               element_j: Tuple[jnp.ndarray, jnp.ndarray]) -> Tuple[jnp.ndarray, jnp.ndarray]:
+    # associative operator for the scan
+    gti, hti = element_i
+    gtj, htj = element_j
+    a = gtj @ gti
+    b = jnp.einsum("...ij,...j->...i", gtj, hti) + htj
+    # clip = 1e9
+    # a = jnp.clip(a, a_min=-clip, a_max=clip)
+    # b = jnp.clip(b, a_min=-clip, a_max=clip)
+    return a, b
+
+
+def matmul_recursive(mats: jnp.ndarray, vecs: jnp.ndarray, y0: jnp.ndarray) -> jnp.ndarray:
+    """
+    Perform the matrix multiplication recursively, y[i + 1] = mats[i] @ y[i] + vec[i].
+
+    Arguments
+    ---------
+    mats: jnp.ndarray
+        The matrices to be multiplied, shape (nsamples - 1, ny, ny)
+    vecs: jnp.ndarray
+        The vector to be multiplied, shape (nsamples - 1, ny)
+    y0: jnp.ndarray
+        The initial condition, shape (ny,)
+
+    Returns
+    -------
+    result: jnp.ndarray
+        The result of the matrix multiplication, shape (nsamples, ny)
+    """
+    # shift the elements by one index
+    eye = jnp.eye(mats.shape[-1], dtype=mats.dtype)[None]  # (1, ny, ny)
+    first_elem = jnp.concatenate((eye, mats), axis=0)  # (nsamples, ny, ny)
+    second_elem = jnp.concatenate((y0[None], vecs), axis=0)  # (nsamples, ny)
+
+    # perform the scan
+    elems = (first_elem, second_elem)
+    _, yt = jax.lax.associative_scan(scan_binop, elems)
+    # jax.debug.print("{nan} {inf}", nan=jnp.any(jnp.isnan(yt)), inf=jnp.any(jnp.isinf(yt)))
+    return yt  # (nsamples, ny)
